Split tests of ResultTypesGenerator into multiple files
